Code style changes

+++ PageResults.cpp
@@ -963,7 +963,7 @@ bool printSearchResultsHeader ( State0 *st ) {
 
 	// save how many docs are in this collection
 	int64_t docsInColl = -1;
-	RdbBase *base = getRdbBase ( (uint8_t)RDB_CLUSTERDB , st->m_collnum );
+	RdbBase *base = getRdbBase ( RDB_CLUSTERDB , st->m_collnum );
 
 	// estimate it
 	if ( base ) {
